requirements.txt
torch~=2.1.0+cu121
typing~=3.7.4.3
numpy~=1.24.4
torchvision~=0.16.0+cu121
pillow~=9.5.0
scikit-learn~=1.3.2
scipy~=1.10.1
opencv-python~=4.8.1.78
matplotlib~=3.7.5
pyyaml~=6.0.1
pandas~=2.0.3
datasets~=3.1.0